+++ MordoEngine/res/shaders/water.fs
@@ -2,12 +2,21 @@
 
 in vec4 ClipSpace;
 
-uniform sampler2D uRefractionTexture;
+uniform sampler2D reflectionTexture;
+uniform sampler2D refractionTexture;
 
 out vec4 FragColor;
 
 void main()
 {
-    vec2 ndc = (ClipSpace.xy / ClipSpace.w) * 0.5 + 0.5;
-    FragColor = texture(uRefractionTexture, ndc);
+    vec2 ndc = ClipSpace.xy / ClipSpace.w;
+    vec2 texCoords = ndc * 0.5 + 0.5;
+
+    vec2 reflectCoords = vec2(texCoords.x, 1.0 - texCoords.y);
+    vec2 refractCoords = texCoords;
+
+    vec4 reflectColor = texture(reflectionTexture, reflectCoords);
+    vec4 refractColor = texture(refractionTexture, texCoords);
+
+    FragColor = mix(reflectColor, refractColor, 0.5);
 }